requirements/base.txt
# Testing
# ------------------------------------------------------------------------------
vcrpy==4.1.1

# Django
# ------------------------------------------------------------------------------
Django==3.1.1
django-environ==0.4.5
djangorestframework==3.12.2
django-health-check==3.16.2

# Utils
# ------------------------------------------------------------------------------
requests==2.24.0
requests-oauthlib==1.3.0
django-memoize==2.3.1
pydantic==1.8.2
